V2XSW-83: SE: Add sync to seClean
To clean SE NVM variables and keystore, the 'vtest clean' command removes
the corresponding filesystem contents, then a board reset is required so
that the hsm clears the keystore contents from its secure memory.  A sync
is recommended before the reset to make sure the fs changes have been
written out.  This change adds the sync operation to the seClean function
so that it does not need to be manually entered at the console before reset

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>