V2XSW-83: LTP integration: return CONF when no test run
Make vtest return CONF when no test has been run.  This avoids counting
a test run as pass in LTP when the test case does not yet exist.  This way
we can define all test cases in LTP, and the ones that have not yet been
implemented will show up as skipped in the test results.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>