V2XSW-143: Vtest clean does not give error if key file does not exist
Modify the vtest clean command so that it does not give an error if the
hsm key file does not exist.  If the key file does not exist, there is no
need to delete it, so add the -f option to the rm command used to delete
the key file (from man rm, the -f option is "ignore nonexistent files,
never prompt").

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>