V2XSW-345 SE: Reduce key generation speed test to 2000 creations (workaround)
That is to make Test 060304 & friends go through without SCU timeout.
The current investigation shows that the issue may come from the SECO
FW, in which dynamic memory is not correctly freed during the key group
retrieval from the NVM.

In order to avoid running into this problem and be able to run the full
vtest test suite, let's implement a temporary workaround which
consists in running less key addition/deletion actions.

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>