V2XSW-333 SE: Update key injection tests for DXL
The comparison after v2xSe_getKek() is now made with all common KEK
values that are known to us (QXP and DXL so far) and returns a valid
test when any of them is matched.

Test patterns for the key injection tests are created based on the
common KEK that is proper to each SoC using a tool script locally.
See previous commit log and tool's usage for details.

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>