V2XSW-62 Use messages as input for sig verif performance test
The current test uses hash values as input for signature verification.
However, the KPI requirement stipulates that messages of max 300 bytes
must be used. In order to take into account the hash computation in the
performance measurements, let's switch the signature verification input
to be messages instead of hash data.

Note that random hash values are still being used for signature
generation tests.

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>